A calculator for speed from delta-distance.
My first little experiment in C++. Literally the first lines of code I wrote. I'm proud of it - I figured out how to calculate an empirical absolute speed using nothing but delta-distance and acceleration. This code will directly drive the UI of my racing game, and some functional bits that will require a direct access to an empirical speed value. 

And for those who are going to ask; since it means I figured out an easy way to calculate deltas in general, acceleration can be calculated from a delta-speed (which is basically a few extra lines of codes and one extra float). PHYSICS!